Added language library support
Tweaked line 408 to call language.GetPhrase before appending text. Native chat.AddText seems to do this by default, or at least similarly. Without this fix, lines may show up as #Prop_Undone instead of "Undone Prop", for example.

I'm not 100% sure this is the "proper" way to fix it, but it seemed to work in my tests.